Tree-Java
The aim of this code is developed a program to add 50.000 records to B
Tree. I generated randomly 50.000 unique numbers for identity key. I
saved identity numbers with name-surname and address information. This
information was generated randomly. The code was completed. Add person
information to B Tree. While adding to B tree, disk-write and disk-read
operation was realized on BTree file. Records was sorted depends on the
identity key number. At the end of program all records was in order in
BTree file.